Fix panel spacer
Summary:
An expanding spacer had it's minimum size set to the config.length
Allowing the panel to go from config size to infinite.

The problem is the panel length effectively defaults to the size the
spacer starts at (it gets set when we drag it into the right place)
meaning it starts off huge and can never shrink. This then leads to
applets being pushed off the panel.

We want it to be in expanding mode from 0 to config.length
and non expanding to ideally be a fixed width - though still probably
compress if it would otherwise mean the applets would overflow.

Setting the preferred size does all that implicitly; and also sets the
appropriate hint so you can have a spacer in a flexible width panel
correctly.

BUG: 352651

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2785