Set explicit minimum size on panelSpacer so that AppletContainer doesn't set one
Summary:
Commit cbeaa821cec7d50a4a2f6cc6f88e0abfa3461b2b in Plasma 5.8 removed a
minimum size on the panel spacer, which is what we want, however if no
minimum size is set on an applet in a panel the panel's AppletContainer
sets the minmimum width to be the same as the height.

This was preventing people having tiny panel spacers.

BUG: 369823

Test Plan: Reporter tested the patch

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2938